+++ src/main/java/plugily/projects/buildbattle/handlers/sign/SignManager.java
@@ -159,14 +159,14 @@ public class SignManager implements Listener {
     if (block == null) {
       return null;
     }
-    ArenaSign arenaSign = null;
+
     for (ArenaSign sign : arenaSigns) {
       if (sign.getSign().getLocation().equals(block.getLocation())) {
-        arenaSign = sign;
-        break;
+        return sign;
       }
     }
-    return arenaSign;
+
+    return null;
   }
 
   public void loadSigns() {